Removed `batch` param and modified tag handling.
Removes the `batch` parameter when creating faxes, and will throw an
error when this parameter is passed in.

Modified how tags are passed. The library user no longer has to manually
write out the tag format given in the Phaxio docs, and can instead
specify an object of `key: value` pairs for tags that are handled
internally.